Increased code for re-download,reNJOY & added log
For all ENDF libraries, EXSAN will detect the presence
of libraries and NJOY processed files. If present,
the user can choose whether to overwrite, clean up (which means
keeping the downloaded zip filel and unpacking it again), or
do nothing.

I also added a log command line option which provides
the capability to redirect stdout to both the screen and a file
called './EXSAN/exsan.log'. This should help with future
debugging as the user base increases.