This improves the mix cross section functionality. Up to now, the mix cross section only displayed micro cross sections weighted by mass fraction or atom ratios of component isotoped.
It is more useful to compute and display macroscopic
cross sections and mean free paths, which requires
computing the number density using the density of
the material in addition to either the mass fraction
or the atom ratio.  This commit adds these features.

The user can now input mixture density (in g/cc) and
if any of the macroscopic cross sections options are
selected, EXSAN will plot them.

I tested it with H2O (using H-1, O-16) and got multigroup
plots that are reasonably consistent with the macroscopic
H2O plots on page 11 in this report:
https://wci.llnl.gov/codes/tart/media/pdf/UCRL-TR-220605.pdf
-HOW 20200118